backend/requirements.txt
meilisearch>=0.31.1
fastapi
uvicorn[standard]
fpdf2
python-jose[cryptography]
passlib[bcrypt]
